+++ benches/bench_compare_rgb16.rs
@@ -79,6 +79,10 @@ pub fn bench_downscale_rgb16(bench: &mut Bench) {
         cpu_ext_and_name.push((CpuExtensions::Sse4_1, "sse4.1"));
         cpu_ext_and_name.push((CpuExtensions::Avx2, "avx2"));
     }
+    #[cfg(target_arch = "aarch64")]
+    {
+        cpu_ext_and_name.push((CpuExtensions::Neon, "neon"));
+    }
     for (cpu_ext, ext_name) in cpu_ext_and_name {
         for alg_name in alg_names {
             let src_view = src_image_data.view();